requirement.txt
matplotlib==3.1.0
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
scipy==1.3.0
torch==1.1.0
torchvision==0.2.2.post3